Fixed WebRTC SDP issue in Webclient with newer versions of Chrome
Newer versions of Chrome cause the audio codecs to look slightly
different in the WebRTC SDP offer that gets sent to the SVMP
daemon. We had code that removed references to the problematic
OPUS codec, however, this code stopped functioning properly
because it was looking for a slightly different substring to remove
from the SDP offer. This caused the WebRTC connection to break and
the video feed to stop working.
Changed this code to be more resilient, now removes OPUS codec info
based on substring anchors (start & end) instead of specifying the
entire substring.